rename MessageStorage to Storage
