+++ src/Beckett/IMessageStore.cs
@@ -1,6 +1,6 @@
 using Beckett.Messages;
-using Beckett.MessageStorage;
 using Beckett.OpenTelemetry;
+using Beckett.Storage;
 
 namespace Beckett;
