+++ src/Beckett/Storage/AppendToStreamResult.cs
@@ -1,3 +1,3 @@
-namespace Beckett.MessageStorage;
+namespace Beckett.Storage;
 
 public readonly record struct AppendToStreamResult(long StreamVersion);
