+++ src/Beckett/Storage/ReadGlobalStreamOptions.cs
@@ -1,4 +1,4 @@
-namespace Beckett.MessageStorage;
+namespace Beckett.Storage;
 
 public class ReadGlobalStreamOptions
 {
