+++ src/Beckett/Storage/ReadStreamResult.cs
@@ -1,3 +1,3 @@
-namespace Beckett.MessageStorage;
+namespace Beckett.Storage;
 
 public record ReadStreamResult(string StreamName, long StreamVersion, IReadOnlyList<StreamMessage> StreamMessages);
