+++ src/Beckett/Subscriptions/GlobalStreamConsumer.cs
@@ -1,7 +1,7 @@
 using System.Threading.Channels;
 using Beckett.Database;
 using Beckett.Database.Types;
-using Beckett.MessageStorage;
+using Beckett.Storage;
 using Beckett.Subscriptions.Queries;
 using Microsoft.Extensions.Logging;
